Fixed some concurrency issues
Move the game id list to it's own object and put some locks in place so
we don't try to read from the map while we're writing to it.

Also add same rate limiting to updating the game id list from steam.
If the last check was less than a half hour before, don't check again.