Add PhotoSwipe gallery to game page
This adds the PhotoSwipe javascript image gallery to the game page.
The gallery needs to have the image sizes before the actual image is
loaded, so they need to be cached.  On startup the RemoteDirectory is
scanned for full size images (not thumbnails) and opens each one to get
the dimensions.  This is a very time consuming process.

It only runs on startup for now.  It will need to be wired into the
directory watcher stuff to work properly.  This gallery is the main
reason for the directory watcher stuff, after all.

Lastly, the metadata needs to cache to disk.  It doesn't do it yet,
which is why it takes forever.