Rewrite image.go; Change refresh strategy
Refactored image.go to not be dumb.  Instead of saving full paths and
using two slices to store image data, just use a map of ImageMeta slices
where the key is the game id.  A bunch of now unused functions were
removed from GameImages as well (like Next(), GetByIndex(), and
Merge()).

Instead of adding a watch to each directory and updating images on the
fly, scan and cache each directory.  Each minute check the cache's
datestamp against the file's ModTime and update when needed.  This
makes subsequent startups an order of magnitude faster.  Startups
without a cache are still long as the cache needs to be built.