Fix directory scanning
Merged the two directory scanning functions and use a goroutine to
refresh the path instead of a timer.

Rewrote a bunch of the image scanning logic.  Turns out it wasn't
adding new files at all on refreshes, but updating modified images
only.  Removed the modified check and make it actually add new files.