Embed static files; rework some HTML/CSS
All files in template/ and static/ are now embedded into the executable
and are served directly from the embedded filesystem.

The unknown.jpg file still needs to be added.

A new staticFiles type has been added that should be able to handle
loading files from disk and storing them in memory.  Idk if this will be
useful at all, but it's there.