Re-add API
The API has been re-added.  The method for adding files has now
completely changed.

The server will only scan the screenshot directory once at start up.
This scan is also blocking and the server will not accept any
connections until the initial scan is done.

When a file is added via the API, the server will save the image to disk
and queue up that filename to be processed in a separate goroutine.
This goroutine will create a thumbnail and add the file to the cache.