Remove RefreshInterval; Update settings example
The refresh interval is no longer used so it was removed.  The server
will only scan the image directory once upon startup.  Connections to
the server will be accepted after this initial scan has completed.

Any images added to the server after startup will need to be done via
the API.